Lookup Barcode on Amazon and Display the Result
When a barcode is scanned, the barcode identifier is assumed
to be an EAN (this assumption may lead to errors for other barcodes)
and an EAN lookup query is made via Amazon's API. The resulting
object or objects are displayed in a list. Clicking on one of the
items in the list will display the web page of the item.